Add platform tracing socket paths and mounts
